Interfaces for Dao. Some interfaces only in progress.
