+++ service/src/main/java/dao/StudentScoreJdbcDao.java
@@ -1,83 +1,48 @@
 package dao;
 
 import model.StudentScore;
-import dao.mappers.StudentScoreMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-
-import java.sql.PreparedStatement;
-import java.sql.Statement;
 import java.util.ArrayList;
 
-@Repository
-public class StudentScoreJdbcDao {
+public interface StudentScoreJdbcDao {
 
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
+    /**
+     * Get StudentScore from source by key : userId and courseId.
+     *
+     * @param userId StudentScore userId in database
+     * @param courseId StudentScore courseId in database
+     * @return StudentScore from database
+     */
+    public StudentScore read(long userId, long courseId);
 
+    /**
+     * Update StudentScore in source.
+     *
+     * @param studentScore for update
+     * @return new update StudentScore
+     */
+    public StudentScore update(StudentScore studentScore);
 
-    public StudentScore read(long userId, long courseId) {
-        String sql = "SELECT * FROM" +
-                " Course_participation" +
-                " student_id = ?," +
-                " course_id = ?";
-        StudentScore studentScore = jdbcTemplate.queryForObject(sql,
-                new StudentScoreMapper(),
-                userId,
-                courseId
-        );
+    /**
+     * Insert StudentScore in data source.
+     *
+     * @param studentScore for insertion in database
+     * @return new StudentScore, which was just now wrote in database
+     */
+    public StudentScore create(StudentScore studentScore);
 
-        return studentScore;
-    }
+    /**
+     * Remove StudentScore from table by id.
+     *
+     * @param userId StudentScore userId
+     * @param courseId StudentScore courseId
+     * @return StudentScore before delete
+     */
+    public StudentScore delete(long userId, long courseId);
 
-    public StudentScore update(StudentScore studentScore) {
-        String sql = "UPDATE Course_participation" +
-                " grade = ?," +
-                " feedback = ?" +
-                "WHERE student_id = ?," +
-                " course_id = ?";
-
-        jdbcTemplate.update(sql,
-                studentScore.getScore(),
-                studentScore.getFeedback(),
-                studentScore.getStudent().getId(),
-                studentScore.getCourse().getId());
-        return studentScore;
-    }
-
-    public StudentScore create(StudentScore studentScore) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        PreparedStatementCreator creator = con -> {
-            PreparedStatement statement =
-                    con.prepareStatement("INSERT INTO " +
-                                    "Course_participation(student_id, " +
-                                    "course_id, " +
-                                    "grade, " +
-                                    "feedback, " +
-                                    "VALUES (?,?,?,?)",
-                            Statement.RETURN_GENERATED_KEYS);
-            statement.setLong(1, studentScore.getCourse().getId());
-            statement.setLong(2, studentScore.getStudent().getId());
-            statement.setInt(3, studentScore.getScore());
-            statement.setString(4, studentScore.getFeedback());
-            return statement;
-        };
-        jdbcTemplate.update(creator, keyHolder);
-        return studentScore;
-    }
-
-    public StudentScore delete(long userId, long courseId) {
-        StudentScore studentScore = read(userId, courseId);
-        String sql = "DELETE FROM Course_participation WHERE student_id = ? and course_id = ?";
-        jdbcTemplate.update(sql, userId, courseId);
-        return studentScore;
-    }
-
-    public ArrayList<StudentScore> list() {
-        return null;
-    }
+    /**
+     * Get list of available StudentScore from database.
+     *
+     * @return ArrayList of available StudentScore
+     */
+    public ArrayList<StudentScore> list();
 }
