+++ service/src/main/java/dao/UserJdbcDao.java
@@ -1,102 +1,54 @@
 package dao;
 
 import model.User;
-import dao.mappers.UserRowMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-import sun.reflect.generics.reflectiveObjects.NotImplementedException;
-
-import java.sql.PreparedStatement;
-import java.sql.Statement;
 import java.util.List;
 
-@Repository
-public class UserJdbcDao {
+public interface UserJdbcDao {
 
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
+    /**
+     * Get User from source by id.
+     *
+     * @param id User id in database
+     * @return User from database, where key is id
+     */
+    public User read(long id);
 
-    public User read(long id) {
-        String sql = "SELECT * FROM User WHERE user_id = ? ";
-        User user = jdbcTemplate.queryForObject(sql,
-                new UserRowMapper(),
-                id
-        );
+    /**
+     * Update User in source.
+     *
+     * @param user for update
+     * @return new update User
+     */
+    public User update(User user);
 
-        return user;
-    }
+    /**
+     * Insert User in data source.
+     *
+     * @param user for insertion in database
+     * @return new User, which was just now wrote in database
+     */
+    public User create(User user);
 
-    public User update(User user) {
-        String sql = "UPDATE User" +
-                " SET" +
-                " first_name = ?," +
-                " last_name = ?," +
-                " login = ?," +
-                " email = ?," +
-                " password = ?" +
-                " WHERE  user_id= ?";
-        jdbcTemplate.update(sql,
-                user.getFirstName(),
-                user.getLastName(),
-                user.getLogin(),
-                user.getEmail(),
-                user.getPassword(),
-                user.getId());
-        return user;
-    }
+    /**
+     * Remove User from table by id.
+     *
+     * @param id User id
+     * @return User before delete
+     */
+    public User delete(long id);
 
-    public User create(User user) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        PreparedStatementCreator creator = con -> {
-            PreparedStatement statement =
-                    con.prepareStatement("INSERT INTO " +
-                                    "User(first_name, " +
-                                    "last_name, " +
-                                    "email, " +
-                                    "login, " +
-                                    "password) " +
-                                    "VALUES (?,?,?,?,?)",
-                            Statement.RETURN_GENERATED_KEYS);
-            statement.setString(1, user.getFirstName());
-            statement.setString(2, user.getLastName());
-            statement.setString(3, user.getEmail());
-            statement.setString(4, user.getLogin());
-            statement.setString(5, user.getPassword());
-            return statement;
-        };
-        jdbcTemplate.update(creator, keyHolder);
-        int id = keyHolder.getKey().intValue();
-        User returnUser = User.newBuilder()
-                .setEmail(user.getEmail())
-                .setFirstName(user.getFirstName())
-                .setLastName(user.getLastName())
-                .setLogin(user.getLogin())
-                .setPassword(user.getPassword())
-                .setId(id)
-                .build();
-        return returnUser;
-    }
+    /**
+     * Get list of available Users from database.
+     *
+     * @return list of available Users
+     */
+    public List<User> list();
 
-    public User delete(long id) {
-        User user = read(id);
-        String sql = "DELETE FROM User WHERE user_id = ?";
-        jdbcTemplate.update(sql, id);
-        return user;
-    }
-
-
-    public List<User> list() {
-        String sql = "SELECT * from User";
-        return  jdbcTemplate.query(sql, new UserRowMapper());
-    }
-
-    public List<User> getStudents(){
-        throw  new NotImplementedException();
-    }
-
+    /**
+     * Get list of available Users from database.
+     *
+     * @return list of Users which are Students
+     */
+    public List<User> getStudents();
 
 }
